AROTCAIS/pyproject.toml
[tool.poetry]
name = "arotcais"
version = "0.1.0"
description = ""
authors = ["SageDesi <sage.desiderio@obf.ateneo.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
